# Garage/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GARAGE - Toda Big Tech tem um inicio",
    "description": "Backend-first engineering education game.",
    "version": "3.0.0"
  },
  "paths": {
...
    "/api/payments/checkout": {
      "post": {
        "tags": [
          "payments"
        ],
        "summary": "Checkout",
        "description": "Generate PIX or card checkout for the selected plan.\n\nDEPRECATED: Now handled entirely by Asaas links on the frontend to avoid creating fake clients dynamically.",
        "operationId": "checkout_api_payments_checkout_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/payments/status/{payment_id}": {
      "get": {
        "tags": [
          "payments"
        ],
        "summary": "Payment Status",
        "description": "Poll Asaas for the payment status.\n\nFrontend polls this every 5 s until status in (CONFIRMED, RECEIVED).\nOn confirmation, it redirects the user to the game.",
        "operationId": "payment_status_api_payments_status__payment_id__get",
        "parameters": [
          {
            "name": "payment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Payment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CheckoutRequest": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id",
            "description": "UUID of the authenticated user"
          },
          "user_name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 2,
            "title": "User Name"
          },
          "user_email": {
            "type": "string",
            "title": "User Email"
          },
          "plan": {
            "type": "string",
            "pattern": "^(monthly|annual)$",
            "title": "Plan"
          },
          "payment_method": {
            "type": "string",
            "pattern": "^(pix|card)$",
            "title": "Payment Method",
            "default": "pix"
          },
          "cpf_cnpj": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cpf Cnpj",
            "description": "Optional CPF/CNPJ for nota fiscal"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "user_name",
          "user_email",
          "plan"
        ],
        "title": "CheckoutRequest"
      },
      "CheckoutResponse": {
        "properties": {
          "payment_id": {
            "type": "string",
            "title": "Payment Id"
          },
          "plan": {
            "type": "string",
            "title": "Plan"
          },
          "payment_method": {
            "type": "string",
            "title": "Payment Method"
          },
          "value": {
            "type": "number",
            "title": "Value"
          },
          "qr_code_base64": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Qr Code Base64"
          },
          "pix_copy_paste": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pix Copy Paste"
          },
          "checkout_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checkout Url"
          },
          "expires_at": {
            "type": "string",
            "title": "Expires At"
          }
        },
        "type": "object",
        "required": [
          "payment_id",
          "plan",
          "payment_method",
          "value",
          "expires_at"
        ],
        "title": "CheckoutResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaymentStatusResponse": {
        "properties": {
          "payment_id": {
            "type": "string",
            "title": "Payment Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "value": {
            "type": "number",
            "title": "Value"
          },
          "plan": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plan"
          }
        },
        "type": "object",
        "required": [
          "payment_id",
          "status",
          "value"
        ],
        "title": "PaymentStatusResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}